No hardcoded secret names
Hardcoding the secret names stops the template from being used for
multiple stacks in the same account.

- Modify the CloudFormation template to use generated names for the
  secrets and set environment variables for the secret ARNs.
- Modify the Lambda to get the secrets based on the ARNs, rather than
  fixed names.

To avoid changing the Lambda code too much, the ARN for any secrets that
are have not been set due to unmatched conditions will set to `Not set`.
This is to keep compatibility with the way the Lambda deals with unset
secrets--it relies on an error when attempting to get the secret value.